requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.26.0
mongomock==4.1.2
pymongo==4.3.3
pytest==7.2.1
python-dotenv==0.19.2
Werkzeug==2.2.3